chore: create story with markdown editor
